+++ CHANGELOG.md
@@ -6,6 +6,7 @@ UNRELEASED
 * [ [#1566](https://github.com/digitalfabrik/integreat-cms/issues/1566) ] Manage organizations per region
 * [ [#1440](https://github.com/digitalfabrik/integreat-cms/issues/1440) ] Add missing word count to translation coverage report
 * [ [#1596](https://github.com/digitalfabrik/integreat-cms/issues/1596) ] Ignore default language in translation report
+* [ [#1607](https://github.com/digitalfabrik/integreat-cms/issues/1607) ] Count words of source translation in translation report
 
 
 2022.8.3